+++ src/index.ts
@@ -1,9 +1,11 @@
 import { Application, RequestHandler } from 'express';
 import { isCelebrate } from 'celebrate';
-import { getSwagger, generateSwagger, writeSwagger } from './swagger';
+import { getSwagger, generateSwagger, writeSwagger, getGlobalSwagger } from './swagger';
 import { validateSchema, validateValue } from './lib/validator';
 import { IOptions, IRouteResult, IExecOptions, ITemplateRoute } from './interfaces';
 import { HttpMethod } from './constants';
+
+const swagger = getGlobalSwagger();
 
 const addRoute = (
   app: Application,
@@ -24,7 +26,7 @@ const addRoute = (
 
   // Autogenerate documentation based on the schema
   if (docs) {
-    generateSwagger(path, method, options);
+    generateSwagger(swagger, path, method, options);
   }
 
   // Add express route (with optional validation)