+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.9.0] - 2023-01-07
+
+### Fixed
+
+- Fixed issue where responses were duplicated in multiple paths in OpenAPI
+- Added tests for swagger generation
+- Added an option to provide initial swagger definition
+
 ## [2.8.0] - 2023-01-06
 
 ### Added